add endpoint for creating business partners in bulk

+++ src/main/kotlin/com/catenax/gpdm/controller/dto/BankAccountDto.kt
@@ -4,7 +4,7 @@ import com.neovisionaries.i18n.CurrencyCode
 import javax.persistence.*
 
 data class BankAccountDto (
-    val trustScores: Collection<Int>,
+    val trustScores: Collection<Float>,
     val currencyCode: CurrencyCode,
     val internationalBankAccountIdentifier: String,
     val internationalBankIdentifier: String,
